+++ crates/nostr-signer/src/lib.rs
@@ -160,7 +160,7 @@ impl NostrSigner {
     {
         let content: &[u8] = content.as_ref();
         match self {
-            Self::Keys(keys) => Ok(nip04::encrypt(keys.secret_key()?, public_key, content)?),
+            Self::Keys(keys) => Ok(nip04::encrypt(keys.secret_key(), public_key, content)?),
             #[cfg(all(feature = "nip07", target_arch = "wasm32"))]
             Self::NIP07(signer) => Ok(signer.nip04_encrypt(public_key, content).await?),
             #[cfg(feature = "nip46")]
@@ -181,7 +181,7 @@ impl NostrSigner {
         let encrypted_content: &str = encrypted_content.as_ref();
         match self {
             Self::Keys(keys) => Ok(nip04::decrypt(
-                keys.secret_key()?,
+                keys.secret_key(),
                 public_key,
                 encrypted_content,
             )?),
@@ -205,7 +205,7 @@ impl NostrSigner {
         let content: &[u8] = content.as_ref();
         match self {
             Self::Keys(keys) => Ok(nip44::encrypt(
-                keys.secret_key()?,
+                keys.secret_key(),
                 public_key,
                 content,
                 nip44::Version::default(),
@@ -229,7 +229,7 @@ impl NostrSigner {
     {
         let payload: &[u8] = payload.as_ref();
         match self {
-            Self::Keys(keys) => Ok(nip44::decrypt(keys.secret_key()?, public_key, payload)?),
+            Self::Keys(keys) => Ok(nip44::decrypt(keys.secret_key(), public_key, payload)?),
             #[cfg(all(feature = "nip07", target_arch = "wasm32"))]
             Self::NIP07(signer) => Ok(signer.nip44_decrypt(public_key, payload).await?),
             #[cfg(feature = "nip46")]