+++ crates/nostr/src/event/builder.rs
@@ -782,7 +782,7 @@ impl EventBuilder {
     ) -> Result<Self, Error> {
         Ok(Self::new(
             Kind::NostrConnect,
-            nip04::encrypt(sender_keys.secret_key()?, &receiver_pubkey, msg.as_json())?,
+            nip04::encrypt(sender_keys.secret_key(), &receiver_pubkey, msg.as_json())?,
             [Tag::public_key(receiver_pubkey)],
         ))
     }
@@ -1288,7 +1288,7 @@ impl EventBuilder {
 
         // Encrypt content
         let content: String = nip44::encrypt(
-            sender_keys.secret_key()?,
+            sender_keys.secret_key(),
             receiver_pubkey,
             rumor.as_json(),
             Version::default(),
@@ -1317,7 +1317,7 @@ impl EventBuilder {
 
         let keys: Keys = Keys::generate();
         let content: String = nip44::encrypt(
-            keys.secret_key()?,
+            keys.secret_key(),
             receiver,
             seal.as_json(),
             Version::default(),